+++ libcxx/include/__type_traits/common_type.h
@@ -25,7 +25,7 @@
 
 _LIBCPP_BEGIN_NAMESPACE_STD
 
-#if __has_builtin(__common_type)
+#if __has_builtin(__builtin_common_type)
 
 template <class... _Args>
 struct common_type;
@@ -34,7 +34,7 @@ template <class... _Args>
 using __common_type_t = typename common_type<_Args...>::type;
 
 template <class... _Args>
-struct common_type : __common_type<__common_type_t, __type_identity, __empty, _Args...> {};
+struct common_type : __builtin_common_type<__common_type_t, __type_identity, __empty, _Args...> {};
 
 #else
 #  if _LIBCPP_STD_VER >= 20